fix: log view refactor for theme
