+++ src/components/DataView/EmptyView.js
@@ -1,0 +1,9 @@
+import { EmptyViewContainer } from "./styled";
+
+export default function EmptyView() {
+    return (
+        <EmptyViewContainer>
+            {"Please adjust search parameters and click on ‘Show Logs’ button"}
+        </EmptyViewContainer>
+    );
+}
