+++ src/components/DataView/consts/consts.js
@@ -1,0 +1,8 @@
+export const TAGS_LEVEL = {
+    critical: ["emerg", "fatal", "alert", "crit", "critical"],
+    error: ["err", "eror", "error", "warning"],
+    warning: ["warn", "warning"],
+    info: ["info", "information", "notice"],
+    debug: ["dbug", "debug"],
+    trace: ["trace"],
+};
