--- src/components/LabelBrowser/LabelBrowser.js
@@ -1,14 +1,0 @@
-import { QueryBar } from "./QueryBar";
-import { ValuesList } from "./ValuesList";
-
-export default function LabelBrowser(){
-
-    return (
-        <div>
-        <QueryBar className={"query-bar-placeholder"}/>
-
-        <ValuesList />
-
-    </div>
-    )
-}
