+++ src/components/LabelBrowser/index.js
@@ -1,0 +1,12 @@
+import { QueryBar } from "./QueryBar";
+import { ValuesList } from "./ValuesList";
+
+export default function LabelBrowser() {
+    return (
+        <div>
+            <QueryBar className={"query-bar-placeholder"} />
+
+            <ValuesList />
+        </div>
+    );
+}
