+++ src/components/StatusBar/actions/setMenuSettingsOpen.js
