+++ src/components/StatusBar/components/daterangepicker/components/Ranges.js
