+++ src/components/StatusBar/components/daterangepicker/components/styles.js
