+++ src/components/StatusBar/components/daterangepicker/defaults.js
