+++ src/plugins/settingsmenu/CloseButton.js
@@ -1,0 +1,10 @@
+import { StyledCloseBtn } from "./styled";
+import CloseIcon from "@mui/icons-material/Close";
+
+export default function CloseButton({ onClick }) {
+    return (
+        <StyledCloseBtn onClick={onClick}>
+            <CloseIcon />
+        </StyledCloseBtn>
+    );
+}
