+++ src/plugins/settingsmenu/CopyButton/CopyButton.js
