+++ src/plugins/settingsmenu/theme/index.js
@@ -1,0 +1,11 @@
+export const SETTINGS_THEME = {
+
+    main:"#ddd",
+    background: "#1a1a1a",
+    header: '#666',
+    inputBg: '#333',
+    buttonBg:'##7b7b7b55',
+    labelColor: '#aaa'    
+}
+
+// import colors from actual theme
